Refactor Consultation Details
